+++ packages/core-js-compat/src/modules-by-versions.js
@@ -55,5 +55,6 @@ module.exports = {
   3.6: [
     'es.regexp.sticky',
     'es.regexp.test',
+    'esnext.array.filter-out',
   ],
 };