+++ tests/pure/index.js
@@ -142,6 +142,7 @@ import './es.weak-set';
 
 QUnit.module('ESNext');
 import './esnext.aggregate-error';
+import './esnext.array.filter-out';
 import './esnext.array.is-template-object';
 import './esnext.async-iterator.constructor';
 import './esnext.async-iterator.as-indexed-pairs';