+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Changelog
+##### Unreleased
+- Added `Array#filterOut` [stage 1 proposal](https://github.com/tc39/proposal-array-filtering)
+
 ##### 3.6.4 - 2020.01.14
 - Prevented a possible almost infinite loop in non-standard implementations of some backward iteration array methods